requirements.txt
streamlit
crewai
langchain
openai
litellm
fastapi
python-dotenv
#litellm[proxy]
apscheduler
pydantic[email]
fastapi-sso